snapshot: Recover inodes and blocks no longer needed
Inodes and blocks are not always freed when files get rewritten or
deleted because these blocks can be part of some snapshot, in which case
they need to be preserved. Whenever snapshots are destroyed by the user,
there is a chance that some inodes or blocks are now not referenced by
any snapshot tree or the active tree. To recover the blocks/inodes, we
iterate through all snapshot inode trees as well as the active inode
tree and set/unset the snapshot and free bitmaps accordingly.

Signed-off-by: Celina Sophie Kalus <celina.kalus@rwth-aachen.de>